Remove flex components, implement Flex and move utilites
